refactor network rule retry
